wallet/requirements.txt
eth_account==0.5.6
web3==5.23.1
bit==0.7.2
python-dotenv==0.19.0
